refactor: Change file package structure
