+++ app/build.gradle.kts
@@ -7,7 +7,7 @@ plugins {
     id("kotlin-parcelize")
     id("kotlin-kapt")
     id("com.google.dagger.hilt.android")
-    //id("com.google.gms.google-services")
+    id("com.google.gms.google-services")
 }
 
 android {
