+++ app/src/main/java/campus/tech/kakao/map/data/SearchResponseDTO.kt
