+++ app/src/main/java/campus/tech/kakao/map/ui/PlaceUiModel.kt
@@ -1,4 +1,6 @@
-package campus.tech.kakao.map.domain
+package campus.tech.kakao.map.ui
+
+import campus.tech.kakao.map.domain.Place
 
 data class PlaceUiModel(
     val placeList: List<Place>,
