+++ build.gradle.kts
@@ -4,7 +4,7 @@ plugins {
     id("org.jetbrains.kotlin.android") version "1.9.0" apply false
     id("org.jlleitschuh.gradle.ktlint") version "12.1.0" apply false
     id("com.google.dagger.hilt.android") version "2.48.1" apply false
-    //id("com.google.gms.google-services") version "4.4.2" apply false
+    id("com.google.gms.google-services") version "4.4.2" apply false
 }
 
 allprojects {
